requirements.txt
asgiref
Django
sqlparse
psycopg2-binary
django-redis
pymongo
PyJWT
djangorestframework